Move torch.Module implementation to norse/torch/module and adjust tests.
